[FLINK-10687] [table] Move format descriptors and validators to flink-table-common
This commit makes the flink-formats module Scala free by introducing a
flink-table-common module that is implemented in Java. This module contains
all classes that are required across different Maven modules.

Additionally, all classes in this module have been annotated with @Internal
and @PublicEvolving accordingly.

Since the methods in Descriptor were declared with private[flink] visibility,
they have been migrated to a new toProperties() method that is public.

This closes #6958.