[FLINK-24405][tests] Introduce util to reliably drain all messages from a kafka topic
The approach used before the introduction was to poll the topic with a
given timeout and if the poll did not return a record treat it as all
records have been read. Unfortunately, due to network issues or similar
the poll can easily return without any records but end offset was not
reached making the tests unreliable. The new util always drains all
messages until the end offset is reached.